remove ChannelCardBase
