+++ src/components/_channel/ChannelCard/ChannelCard.styles.ts
