--- src/components/_channel/ChannelCardBase/index.ts
@@ -1,1 +1,0 @@
-export * from './ChannelCardBase'
